requirements.txt
google-cloud-speech==2.34.0
watchdog==6.0.0
websockets==15.0.1
python-dotenv==1.2.1
